requirements.txt
streamlit>=1.24.0
pandas>=1.3.0
numpy>=1.21.0
faiss-cpu
sentence-transformers
transformers
scikit-learn>=1.0.0
joblib
